+++ lib/src/cook/Context.cpp
@@ -2,8 +2,6 @@
 #include "cook/algo/Book.hpp"
 #include "cook/generator/graphviz/Dependency.hpp"
 #include "cook/generator/graphviz/Component.hpp"
-#include "cook/generator/qt/Generic.hpp"
-#include "cook/generator/qt/Cook.hpp"
 #include "cook/generator/CMake.hpp"
 #include "cook/generator/Naft.hpp"
 #include "cook/generator/Ninja.hpp"
@@ -38,22 +36,10 @@ Result Context::initialize()
     // add the generators
     MSS(register_generator(std::make_shared<generator::graphviz::Dependency>()));
     MSS(register_generator(std::make_shared<generator::graphviz::Component>()));
-    MSS(register_generator(std::make_shared<generator::qt::Generic>()));
-    MSS(register_generator(std::make_shared<generator::qt::Cook>()));
     MSS(register_generator(std::make_shared<generator::CMake>()));
     MSS(register_generator(std::make_shared<generator::Naft>()));
     MSS(register_generator(std::make_shared<generator::Ninja>()));
     MSS(register_generator(std::make_shared<generator::HTML>()));
-
-    // specify as project name the current directory
-    {
-        std::filesystem::path fn = gubg::filesystem::combine(std::filesystem::current_path(), dirs().recipe());
-        if (!fn.empty())
-        {
-            auto it = --fn.end();
-            set_project_name(it->string());
-        }
-    }
 
     MSS_END();
 }