+++ changelog.md
@@ -94,8 +94,7 @@ Proof-of-concept version.
 * Added Framework and FrameworkDir
 * temporary output files are under <output_dir> / <recipe> / <md5 of dir> / <rel>
 * Comparing chai recipes and books
-* [Project|Output|Temporary] dir are exposed in chai (read-only)
-* Combining absolute and relative path works
+* [Current|Output|Temporary] dir are exposed in chai (read-only)
 * Flags are clonable (copy constructable)
 * [wip] Support for HTML output